+++ internal/pkg/node/flags.go
@@ -340,4 +340,3 @@ func createUnsetFlag(baseCmd *cobra.Command, myType reflect.StructField, unsetMa
 		}
 	}
 }
-